requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
pydantic-settings==2.1.0
redis==5.0.1
python-multipart==0.0.6
httpx==0.26.0
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
email-validator==2.1.0.post1
argon2-cffi==21.3.0
